+++ src/api/json-api/index.ts
@@ -1,3 +1,0 @@
-export { JsonApiClient } from './json-api'
-export { JsonApiResponse } from './response'
-export * from './errors'